refactor(spring-boot): introduce FormFieldsManager bean
Encapsulate form-field logic (session prefix + next session name) in a
FormFieldsManager bean, injected into JkleeFilesEndpoint. FormFields is now
a pure value object. The bean resolves the session-prefix fallback to
spring.application.name in auto-config.
